Added Image Loading, Optimized Rendering, Added Input System, Added stb_image dependency
- Added image loading, supports byte arrays and files with stb_image

- Optimized rendering loop by uploading all shapes at once, then drawindexedinstanced batches of shapes

- Added inputsystem process->input

+++ flashgui/fonts.h
@@ -18,8 +18,8 @@ namespace fgui {
     struct font_glyph_info {
         float u0, v0, u1, v1; // UV rect in atlas
         float advance;        // in pixels (floating to allow sub-pixel placement)
-        int offset_x;         // horizontal offset (pixels) from pen position to bitmap origin
-        int offset_y;         // vertical offset (pixels) from baseline to bitmap origin
+        float offset_x;         // horizontal offset (pixels) from pen position to bitmap origin
+        float offset_y;         // vertical offset (pixels) from baseline to bitmap origin
         DWRITE_GLYPH_METRICS metrics; // glyph metrics from DirectWrite
     };
